+++ access-service/src/test/java/cn/ac/fage/accessmesh/access/resource/service/impl/ResourceBatchCreateCompositeIdentityPgIT.java
@@ -106,6 +106,7 @@ class ResourceBatchCreateCompositeIdentityPgIT {
 
     private ResourceManageAppServiceImpl newResourceManageAppService(PermQueryEngine engine) {
         return new ResourceManageAppServiceImpl(resourceEntityMapper, resourceApiMappingMapper,
+            org.mockito.Mockito.mock(cn.ac.fage.accessmesh.access.resource.mapper.ServiceConfigMapper.class),
             resourceEntityDomainService, typeResolutionService, domainClassifyService,
             engine, rolePermDomainService,
             mock(cn.ac.fage.accessmesh.access.rule.service.domain.PermissionConditionDomainService.class),