+++ access-service/src/test/java/cn/ac/fage/accessmesh/access/resource/service/impl/ServiceConfigCascadePgIT.java
@@ -125,6 +125,7 @@ class ServiceConfigCascadePgIT {
 
     private ResourceManageAppService newResourceManageAppService(PermQueryEngine engine) {
         return new ResourceManageAppServiceImpl(resourceEntityMapper, resourceApiMappingMapper,
+            org.mockito.Mockito.mock(cn.ac.fage.accessmesh.access.resource.mapper.ServiceConfigMapper.class),
             resourceEntityDomainService, typeResolutionService, domainClassifyService,
             engine, rolePermMapper, org.mockito.Mockito.mock(cn.ac.fage.accessmesh.access.rule.service.domain.PermissionConditionDomainService.class), ownershipGuard(),
             mock(TreeWriteLockSupport.class));